e2e: move optional operators into their own package
We also now require the `optional_operators` build tag to build them.
The e2e make targets pass this by default for now, so no change to the
overall list of tests has been made.
